+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 2022-04-29
+- Fixed Focus hierarchy to selection falling back to the scene view hierarchy when an asset is focused in the Inspector.
+- Fixed Focus hierarchy to selection not repainting unfocused windows.
+
 ## [1.1.1] - 2022-04-17
 - Added Focus hierarchy to selection (Alt+Shift+F). This is likely more useful than Collapse Hierarchy. It supports:
   - Hierarchy view